+++ fiddles/bash/fiddle-0057-PhantomJsRasterize/script.sh
@@ -1,0 +1,11 @@
+#!/usr/bin/env bash
+
+clear;
+echo "$0" | sed 's/\.\///g' | awk '{print toupper($0)}'
+echo "Bash version ${BASH_VERSION}..."
+
+function screenshot() {
+   phantomjs screenshot.js "http://localhost:5555/angular2/fiddle-0001-HelloWorld/index.html" screenshot.png
+}
+
+screenshot;